Rename user pkg to broker
As the other pkgs were separated to its own library it makes more sense
to call the `user` pkg `broker` instead. This also avoids the need for
the consumer to later rename the `user` pkg, as the name is generic
enough that it can be kept.
