+++ cmd/servicebroker/main.go
@@ -16,11 +16,11 @@ import (
 	"github.com/pmorie/osb-broker-lib/pkg/metrics"
 	"github.com/pmorie/osb-broker-lib/pkg/rest"
 	"github.com/pmorie/osb-broker-lib/pkg/server"
-	"github.com/pmorie/osb-starter-pack/pkg/user"
+	"github.com/pmorie/osb-starter-pack/pkg/broker"
 )
 
 var options struct {
-	user.Options
+	broker.Options
 
 	Port    int
 	TLSCert string
@@ -31,7 +31,7 @@ func init() {
 	flag.IntVar(&options.Port, "port", 8005, "use '--port' option to specify the port for broker to listen on")
 	flag.StringVar(&options.TLSCert, "tlsCert", "", "base-64 encoded PEM block to use as the certificate for TLS. If '--tlsCert' is used, then '--tlsKey' must also be used. If '--tlsCert' is not used, then TLS will not be used.")
 	flag.StringVar(&options.TLSKey, "tlsKey", "", "base-64 encoded PEM block to use as the private key matching the TLS certificate. If '--tlsKey' is used, then '--tlsCert' must also be used")
-	user.AddUserFlags(&options.Options)
+	broker.AddFlags(&options.Options)
 	flag.Parse()
 }
 
@@ -62,7 +62,7 @@ func runWithContext(ctx context.Context) error {
 
 	addr := ":" + strconv.Itoa(options.Port)
 
-	businessLogic, err := user.NewBusinessLogic(options.Options)
+	businessLogic, err := broker.NewBusinessLogic(options.Options)
 	if err != nil {
 		return err
 	}
