+++ pkg/broker/doc.go
@@ -1,5 +1,5 @@
-// Package user holds the code that users of the skeleton write for their
-// broker.  To make a broker, fill out:
+// Package broker holds the code that users of the skeleton write for their
+// broker. To make a broker, fill out:
 //
 // - The Options type, which holds options for the broker
 // - The AddFlags function, which adds CLI flags for an Options
@@ -7,4 +7,4 @@
 //   business logic
 // - The NewBusinessLogic function, which creates a BusinessLogic from the
 //   Options the program is run with
-package user // import "github.com/pmorie/osb-starter-pack/pkg/user"
+package broker // import "github.com/pmorie/osb-starter-pack/pkg/broker"
